% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarbonateEq.R
\name{CarbonateEq}
\alias{CarbonateEq}
\title{R6 class defining carbonate equilibrium}
\description{
Defines a set of equilibrium
  conditions for inorganic carbon at a given temperature 
  in freshwater.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tempC}}{Temperature in deg Celsius}

\item{\code{tempK}}{Temperature in Kelvin}

\item{\code{eConduct}}{Electrical conductivity}

\item{\code{ionicStrength}}{Ionic strength}

\item{\code{daviesParam}}{Parameter of Davies equation}

\item{\code{daviesExponent}}{Exponent of Davies equation}

\item{\code{activityCoeffH}}{Activity coefficient for hydrogen ions}

\item{\code{activityCoeffOH}}{Activity coefficient for hydroxide ions}

\item{\code{activityCoeffHCO3}}{Activity coefficient for bicarbonate ions}

\item{\code{activityCoeffCO3}}{Activity coefficient for carbonate ions}

\item{\code{kDissocH2CO3App}}{Apparent equilibrium constant for carbonic acid disassociation}

\item{\code{kDissocHCO3App}}{Apparent equilibrium constant for bicarbonate disassociation}

\item{\code{kDissocH2OApp}}{Apparent equilibrium constant for water disassociation}

\item{\code{kHenryCO2}}{Henry's constant}

\item{\code{kHenryCO2fromTempFunc}}{Function used to calculate Henry's constant}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CarbonateEq$new()}}
\item \href{#method-resetFromTemp}{\code{CarbonateEq$resetFromTemp()}}
\item \href{#method-calcTotalAlkFromDICpH}{\code{CarbonateEq$calcTotalAlkFromDICpH()}}
\item \href{#method-calcTotalAlkFromConcCO2pH}{\code{CarbonateEq$calcTotalAlkFromConcCO2pH()}}
\item \href{#method-calcDICFromConcCO2pH}{\code{CarbonateEq$calcDICFromConcCO2pH()}}
\item \href{#method-optpHFromDICTotalAlk}{\code{CarbonateEq$optpHFromDICTotalAlk()}}
\item \href{#method-optpHFromConcCO2TotalAlk}{\code{CarbonateEq$optpHFromConcCO2TotalAlk()}}
\item \href{#method-optpHFromConcCO2DIC}{\code{CarbonateEq$optpHFromConcCO2DIC()}}
\item \href{#method-optfCO2FromDICTotalAlk}{\code{CarbonateEq$optfCO2FromDICTotalAlk()}}
\item \href{#method-optDICFromfCO2TotalAlk}{\code{CarbonateEq$optDICFromfCO2TotalAlk()}}
\item \href{#method-optTotalAlkFromfCO2DIC}{\code{CarbonateEq$optTotalAlkFromfCO2DIC()}}
\item \href{#method-clone}{\code{CarbonateEq$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct an instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$new(tempC, kHenryCO2fromTempFunc = kHenryCO2fromTemp, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tempC}}{Water temperature in degrees Celsius}

\item{\code{kHenryCO2fromTempFunc}}{Optional function to use for calculating Henry's constant.
Defaults to kHenryCO2fromTemp provided with the package.}

\item{\code{...}}{Additional arguments will be 
passed generically to the method \code{resetFromTemp()} 
when it is called to initialize the object. 
See documentation for \code{resetFromTemp()}
for description of arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resetFromTemp"></a>}}
\if{latex}{\out{\hypertarget{method-resetFromTemp}{}}}
\subsection{Method \code{resetFromTemp()}}{
Resets equilibrium constants in attributes of the object
   based on the provided temperature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$resetFromTemp(
  tempC,
  tempK = tempC + 273.15,
  eConduct = 0,
  ionicStrength = 0.013 * eConduct,
  daviesParam = 0.5092 + (tempC - 25) * 0.00085,
  daviesExponent = -daviesParam * (ionicStrength^0.5/(1 + ionicStrength^0.5) - 0.3 *
    ionicStrength)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tempC}}{Water temperature in degrees Celsius}

\item{\code{tempK}}{Temperature in Kelvin. Defaults to tempC argument plus 273.15}

\item{\code{eConduct}}{Electrical conductivity in milliSiemens per cm, default is set to zero
which will result in the ideal case of solute activities equal to 
their concentrations}

\item{\code{ionicStrength}}{Ionic strength in moles per liter, default is set as a linear correlate
of electrical conductivity as suggested by 
Griffin and Jurinak (1973 - Soil Science)}

\item{\code{daviesParam}}{Parameter for Davies equation derivation, defaults to
\code{0.5092 + (tempC - 25) * 0.00085}}

\item{\code{daviesExponent}}{Factor for exponent in Davies equation, defaults to
\code{(ionicStrength ^ 0.5 / (1 + ionicStrength ^ 0.5) - 0.3 * ionicStrength)}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named vector with 2 values: the previous temperature and the new temperature
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calcTotalAlkFromDICpH"></a>}}
\if{latex}{\out{\hypertarget{method-calcTotalAlkFromDICpH}{}}}
\subsection{Method \code{calcTotalAlkFromDICpH()}}{
Calculates the total alkalinity based on a provided DIC concentration
  and pH. This is a simplified alkalinity based only on carbonate species.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$calcTotalAlkFromDICpH(concDIC, pH)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concDIC}}{Dissolved inorganic carbon concentration in molality}

\item{\code{pH}}{The pH of the solution (minus the log10 of H+ concentration in molality)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Total alkalinity in charge molality
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calcTotalAlkFromConcCO2pH"></a>}}
\if{latex}{\out{\hypertarget{method-calcTotalAlkFromConcCO2pH}{}}}
\subsection{Method \code{calcTotalAlkFromConcCO2pH()}}{
Calculates the total alkalinity based on a provided carbon dioxide concentration
  and pH. This is a simplified alkalinity based only on carbonate species.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$calcTotalAlkFromConcCO2pH(concCO2, pH)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concCO2}}{Concentration of carbon dioxide in molality}

\item{\code{pH}}{The pH of the solution (minus the log10 of H+ concentration in molality)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Total alkalinity in charge molality
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calcDICFromConcCO2pH"></a>}}
\if{latex}{\out{\hypertarget{method-calcDICFromConcCO2pH}{}}}
\subsection{Method \code{calcDICFromConcCO2pH()}}{
Calculates dissolved inorganic carbon based on a provided carbon dioxide concentration
  and pH.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$calcDICFromConcCO2pH(concCO2, pH)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concCO2}}{Concentration of carbon dioxide in molality}

\item{\code{pH}}{The pH of the solution (minus the log10 of H+ concentration in molality)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Dissolved inorganic carbon concentration in molality of C
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optpHFromDICTotalAlk"></a>}}
\if{latex}{\out{\hypertarget{method-optpHFromDICTotalAlk}{}}}
\subsection{Method \code{optpHFromDICTotalAlk()}}{
Uses a single parameter optimization to infer the pH from a given
   DIC concentration and total alkalinity. The pH is inferred by 
   changing its value until the resulting calculated alkalinity
   matches the provided alkalinity.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$optpHFromDICTotalAlk(
  concDIC,
  totalAlk,
  tolerance = 1e-05,
  range = c(12, 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concDIC}}{Dissolved inorganic carbon concentration in molality}

\item{\code{totalAlk}}{The total alkalinity in charge molality}

\item{\code{tolerance}}{The tolerance used for convergence in the optimization algorithm.
Default value is 1e-5.}

\item{\code{range}}{Range in pH values to constrain the optimization.
Default range is 2 < pH < 12.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The inferred pH.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optpHFromConcCO2TotalAlk"></a>}}
\if{latex}{\out{\hypertarget{method-optpHFromConcCO2TotalAlk}{}}}
\subsection{Method \code{optpHFromConcCO2TotalAlk()}}{
Uses a single parameter optimization to infer the pH from a given
   carbon dioxide concentration and total alkalinity. The pH is inferred by 
   changing its value until the resulting calculated alkalinity
   matches the provided alkalinity.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$optpHFromConcCO2TotalAlk(
  concCO2,
  totalAlk,
  tolerance = 1e-05,
  range = c(12, 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concCO2}}{Dissolved carbon dioxide concentration in molality}

\item{\code{totalAlk}}{The total alkalinity in charge molality}

\item{\code{tolerance}}{The tolerance used for convergence in the optimization algorithm.
Default value is 1e-5.}

\item{\code{range}}{Range in pH values to constrain the optimization.
Default range is 2 < pH < 12.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The inferred pH.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optpHFromConcCO2DIC"></a>}}
\if{latex}{\out{\hypertarget{method-optpHFromConcCO2DIC}{}}}
\subsection{Method \code{optpHFromConcCO2DIC()}}{
Uses a single parameter optimization to infer the pH from provided
   carbon dioxide and dissolved inorganic carbon concentrations . 
   The pH is inferred by 
   changing its value until the resulting calculated dissolved inorganic
   carbon matches the provided value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$optpHFromConcCO2DIC(
  concCO2,
  concDIC,
  tolerance = 1e-05,
  range = c(12, 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concCO2}}{Dissolved carbon dioxide concentration in molality}

\item{\code{concDIC}}{The dissolved inorganic carbon in molality of C}

\item{\code{tolerance}}{The tolerance used for convergence in the optimization algorithm.
Default value is 1e-5.}

\item{\code{range}}{Range in pH values to constrain the optimization.
Default range is 2 < pH < 12.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The inferred pH.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optfCO2FromDICTotalAlk"></a>}}
\if{latex}{\out{\hypertarget{method-optfCO2FromDICTotalAlk}{}}}
\subsection{Method \code{optfCO2FromDICTotalAlk()}}{
Uses a pH optimization to infer the fugacity of carbon dioxide based
   on the provided DIC concentration and total alkalinity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$optfCO2FromDICTotalAlk(
  concDIC,
  totalAlk,
  tolerance = 1e-05,
  range = c(12, 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{concDIC}}{Dissolved inorganic carbon concentration in molality}

\item{\code{totalAlk}}{The measured or otherwise known total alkalinity in charge molality}

\item{\code{tolerance}}{The tolerance used for convergence in the optimization algorithm,
default value is 1e-5}

\item{\code{range}}{Range in pH values to constrain the optimization,
default range is 2 < pH < 12}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of the inferred pCO2 and pH.
   Value of pCO2 is in units of microatmospheres.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optDICFromfCO2TotalAlk"></a>}}
\if{latex}{\out{\hypertarget{method-optDICFromfCO2TotalAlk}{}}}
\subsection{Method \code{optDICFromfCO2TotalAlk()}}{
Uses a pH optimization to infer the dissolved inorganic carbon based
   on the provided fugacity of carbon dioxide and total alkalinity
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$optDICFromfCO2TotalAlk(
  fCO2,
  totalAlk,
  tolerance = 1e-05,
  range = c(12, 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fCO2}}{Fugacity of carbon dioxide in microatmospheres}

\item{\code{totalAlk}}{The measured or otherwise known total alkalinity in charge molality}

\item{\code{tolerance}}{The tolerance used for convergence in the optimization algorithm,
default value is 1e-5}

\item{\code{range}}{Range in pH values to constrain the optimization,
default range is 2 < pH < 12}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of the inferred DIC concentration in molality and pH.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optTotalAlkFromfCO2DIC"></a>}}
\if{latex}{\out{\hypertarget{method-optTotalAlkFromfCO2DIC}{}}}
\subsection{Method \code{optTotalAlkFromfCO2DIC()}}{
Uses a pH optimization to infer the total alkalinity based
   on the provided fugacity of carbon dioxide and dissolved
   inorganic carbon. Total alkalinity is assumed to be dominated
   by carbonate species.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$optTotalAlkFromfCO2DIC(
  fCO2,
  concDIC,
  tolerance = 1e-05,
  range = c(12, 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fCO2}}{Fugacity of carbon dioxide in microatmospheres}

\item{\code{concDIC}}{The dissolved inorganic carbon in molality of C}

\item{\code{tolerance}}{The tolerance used for convergence in the optimization algorithm,
default value is 1e-5}

\item{\code{range}}{Range in pH values to constrain the optimization,
default range is 2 < pH < 12}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named list of the inferred alkalinity in charge molality and pH.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CarbonateEq$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
