% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsUtilities.R
\name{densityWater}
\alias{densityWater}
\title{Calculates the density of water}
\usage{
densityWater(temp)
}
\arguments{
\item{temp}{Numerical vector of temperatures (deg C)}
}
\value{
Numerical vector of densities of water in kilograms per liter
}
\description{
Density of water is calculated based on an empirical 
   relationship with the temperature of water
}
