% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsUtilities.R
\name{DoSatCalculator}
\alias{DoSatCalculator}
\title{R6 class defining a DO Saturation concentration calculator}
\description{
Calculates the saturation concentration of dissolved
   oxygen in water based on temperature and air pressure.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{densityWaterFunc}}{The function used to calculate the density of water from temperature}

\item{\code{stdAirPressure}}{The air pressure under standard conditions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DoSatCalculator$new()}}
\item \href{#method-calculate}{\code{DoSatCalculator$calculate()}}
\item \href{#method-clone}{\code{DoSatCalculator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct an instance of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DoSatCalculator$new(densityWaterFunc = densityWater, stdAirPressure = 760)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{densityWaterFunc}}{The function used to estimate the density of water
from temperature.  Defaults to densityWater provided as
a utility function in this package.}

\item{\code{stdAirPressure}}{Standard air pressure at sea level in the same units
that will be used for the saturation concentration
calculation.  Defaults to 760 mm Hg. Units here
must be consistent with units used in other methods.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calculate"></a>}}
\if{latex}{\out{\hypertarget{method-calculate}{}}}
\subsection{Method \code{calculate()}}{
Calculates the saturated dissolved oxygen concentration based
  on the provided water temperature and air pressure.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DoSatCalculator$calculate(temp, airPressure)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{temp}}{Water temperature(s) in degrees Celsius}

\item{\code{airPressure}}{Air pressure in same units as the definition of standard
air pressure in the class attribute stdAirPressure}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The dissolved oxygen concentration at saturation relative
  to concentrations in the overlying atmosphere (Henry's law)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DoSatCalculator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
