\name{exampleHtest}
\alias{exampleHtest}
\docType{data}
\title{
An example of the results obtained using the function htest.
}
\description{
The result from htest is a data frame containing the metabolites present in each biological sample, their respective intensities across samples and their associated p-values resultant from the statistical analysis performed.
}
\usage{data(exampleHtest)}
\format{
  A data frame consisting of metabolites and their intensities/abundances in each biological sample.
}
\examples{
data(exampleHtest)
}
\keyword{
htest
}