\name{exampleMetReport}
\alias{exampleMetReport}
\docType{data}
\title{
An example of the results obtained using the function exampleMetReport.
}
\description{
The result from exampleMetReport is a data frame containing the metabolites present in each biological sample with their respective intensities across samples.
}
\usage{data(exampleMetReport)}
\format{
  A data frame consisting of metabolites and their intensities/abundances in each biological sample.
}
\examples{
data(exampleMetReport)
}
\keyword{MetReport}