\name{exampleMSLfile}
\alias{exampleMSLfile}
\docType{data}
\title{
An example of MSL file of an AMDIS library.
}
\description{
The MSL file of an AMDIS library is a text file containing the spectrum and RT of each metabolite in the spectral library.
}
\usage{data(exampleMSLfile)}
\format{
  A data frame.
  \describe{
    \item{\code{V1}}{Example of the MSL file of an AMDIS library}
  }
}
\examples{
data(exampleMSLfile)
print(exampleMSLfile)
}
\keyword{datasets}