% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarbonateEq.R
\name{kHenryCO2fromTemp}
\alias{kHenryCO2fromTemp}
\title{Calculate Henry's constant for carbon dioxide}
\usage{
kHenryCO2fromTemp(tempK)
}
\arguments{
\item{tempK}{Temperature in Kelvin}
}
\value{
Henry's constant for CO2 in moles per liter per atmosphere
}
\description{
Calculates Henry's constant for carbon dioxide dissolution in 
   water based on the provided temperature. 
   (Weiss 1974 - Marine Chemistry)
}
