% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsUtilities.R
\name{kSchmidtCO2}
\alias{kSchmidtCO2}
\title{Adjusts air-freshwater dissolved carbon dioxide exchange for temperature}
\usage{
kSchmidtCO2(temp, k600)
}
\arguments{
\item{temp}{Water temperature in degrees Celsius
(numerical vector)}

\item{k600}{Gas exchange rate (per time) or velocity (length per time)
at a Schmidt number of 600
(numerical vector)}
}
\value{
Numerical vector of temperature corrected gas exchange rates 
   for carbon dioxide in same units as k600 argument
}
\description{
Adjusts a dissolved carbon dioxide exchange velocity or rate based on temperature
   and the provided dissolved gas exchange velocity or rate at a Schmidt number of 600.
   The Schmidt number for carbon dioxide is calculated based on an empirical function of 
   temperature accoring to Raymond et al. (2012 Limnology and Oceanography: Fluids and Environments).
   The exchange rate at a Schmidt number of 600 is then coverted to the
   gas exchange rate at the Schmidt number estimated from the provided temperature
   according to Jahne et al. (1987, Journal of Geophysical Research).
}
