\name{Metab-package}
\alias{Metab-package}
\alias{Metab}
\docType{package}
\title{
Metab processes metabolomics data previously analyzed by the Automated Mass Spectral and Deconvolution System (AMDIS).
}
\description{
Metab is an R package for high-throughput processing of metabolomics data analysed by the Automated Mass Spectral Deconvolution and Identification System (AMDIS) (http://chemdata.nist.gov/mass-spc/amdis/downloads/). In addition, it performs statistical hypothesis test (t-test) and analysis of variance (ANOVA). Doing so, Metab considerably speed up the data mining process in metabolomics and produces better quality results. Metab was developed using interactive features, allowing users with lack of R knowledge to appreciate its functionalities.
}
\details{
\tabular{ll}{
Package: \tab Metab\cr
Type: \tab Package\cr
Version: \tab 0.99.6\cr
Date: \tab 2013-10-11\cr
License: \tab GNU General Public License as published by the Free\\
 Software Foundation, either version 3 of the License, or any later version.\cr
}
}
\author{
Raphael Aggio

Maintainer: Raphael Aggio <ragg005@aucklanduni.ac.nz>
}
\references{
Aggio, R., Villas-Boas, S. G., & Ruggiero, K. (2011). Metab: an R package for high-throughput analysis of metabolomics data generated by GC-MS. Bioinformatics, 27(16), 2316-2318. doi: 10.1093/bioinformatics/btr379
}
\keyword{ Metabolomics }
\seealso{
\code{\link{htest}}, 
\code{\link{MetReport}}, 
\code{\link{MetReportNames}}, 
\code{\link{normalizeByInternalStandard}}, 
\code{\link{removeFalsePositives}}, 
\code{\link{buildLib}},
\code{\link{normalizeByBiomass}} 
}
\examples{
##### Load example data #####
data(exampleMetReport)
##### print data ############
print(exampleMetReport) 
}
