\name{normalizeByBiomass}
\alias{normalizeByBiomass}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to normalize metabolomics data by the biomass associated to each biological sample (e.g. cell number or O.D.)
}
\description{
normalizeByBiomass divides the intensity of each metabolite in a specific sample by the value of the biomass measured for this specific sample. 
}
\usage{
normalizeByBiomass(inputData, biomass, save = TRUE, folder, output = "norm_bio") 
}
\arguments{
  \item{inputData}{
When inputData is missing, a dialog box will pop up allowing the user to click-and-point to the .csv file from which the data is to be read. It may also receive a character string pointing to a .csv file containing a data frame such as data(exampleMetReport), generated by \code{\link{MetReport}}. Alternatively, inputData takes an R vector containing the desired data frame.
}
  \item{biomass}{
When biomass is missing, a dialog box will pop up allowing the user to click-and-point to the .csv file from which the biomasses are to be read (See data(exampleBiomass)). It may also receive a character string pointing to a .csv file containing a data frame such as data(exampleBiomass). Alternativelly, it may receive a character vector indicating the data frame such as data(exampleBiomass). 
}
  \item{save}{
A logical vector (TRUE or FALSE) indicating if the resultant data frame should be saved in a .csv file. If save = TRUE, the .csv file will be saved in the path defined in the argument folder.
}
  \item{folder}{
A character vector indicating the path to the folder where the results will be saved.
}
  \item{output}{
A character vector indicating the name of the .csv file to be generated.
}
}
\details{
normalizeByBiomass loads the inputData and searches the data frame Biomass for the values of the biomasses associated with each sample present in the inputData. When the value is found, the abundances of metabolites associated with each sample is divided by the value of its respective biomass. 	
	
}
\value{
norm.biomass produces a data frame normalized by biomass.

Note that the first line of the resulting data.frame is used to represent sample meta-data (for example replicates). 
}
\references{
Aggio, R., Villas-Boas, S. G., & Ruggiero, K. (2011). Metab: an R package for high-throughput analysis of metabolomics data generated by GC-MS. Bioinformatics, 27(16), 2316-2318. doi: 10.1093/bioinformatics/btr379
}
\author{
Raphael Aggio <ragg005@aucklanduni.ac.nz>
}
\seealso{
\code{\link{htest}}, 
\code{\link{MetReport}}, 
\code{\link{MetReportNames}}, 
\code{\link{normalizeByInternalStandard}}, 
\code{\link{removeFalsePositives}}, 
\code{\link{buildLib}} 
}
\examples{
### Load the inputData ###
data(exampleMetReport)
### Load the list of biomasses ### 
data(exampleBiomass)
### Normalize ####
normalizedData <- normalizeByBiomass(exampleMetReport, biomass = exampleBiomass, save = FALSE)
### Show results ###
print(normalizedData)
}