\name{normalizeByInternalStandard}
\alias{normalizeByInternalStandard}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Normalize metabolomics data sets by a specific metabolite defined as internal standard.
}
\description{
In the specified inputData, every metabolite from each sample will be divided by the intensity/abundance of the metabolite defined as internal standard, which is specified through the argument internalStandard.
}
\usage{
normalizeByInternalStandard(
	inputData, 
	internalStandard, 
	save = TRUE, 
	folder, 
	output = "normalizedByInternalStandard"
)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{inputData}{
When inputData is missing, a dialog box will pop up allowing the user to click-and-point to the .csv file from which the data is to be read. It may also receive a character string pointing to a .csv file containing a data frame such as data(exampleMetReport), generated by \code{\link{MetReport}}. Alternatively, inputData takes an R vector containing the desired data frame.
}
  \item{internalStandard}{
A character string indicating the name of the compound to be used as internal standard. If internalStandard is missing, a list of metabolites is presented to the user to interactively choose the correct compound.
}
  \item{save}{
A logical vector (TRUE or FALSE) indicating if the resultant data frame should be saved in a .csv file. If save = TRUE, the .csv file will be saved in the path defined in the argument folder.
}
  \item{folder}{
A character string indicating the path to the folder where the results will be saved.
}
  \item{output}{
A character string indicating the name of the .csv file to be generated.
}
}
\details{
normalizeByInternalStandard will divide the abundances of each metabolite in a specific sample by the abundance of the chosen internal standard in this specific sample.
}
\value{
normalizeByInternalStandard generates a data frame containing metabolite abundances normalized by the nominated internal standard.

Note that the first line of the resulting data.frame is used to represent sample meta-data (for example replicates). 

}
\references{
Aggio, R., Villas-Boas, S. G., & Ruggiero, K. (2011). Metab: an R package for high-throughput analysis of metabolomics data generated by GC-MS. Bioinformatics, 27(16), 2316-2318. doi: 10.1093/bioinformatics/btr379
}
\author{
Raphael Aggio <ragg005@aucklanduni.ac.nz>
}
\seealso{
\code{\link{htest}}, 
\code{\link{MetReport}}, 
\code{\link{MetReportNames}}, 
\code{\link{normalizeByBiomass}}, 
\code{\link{removeFalsePositives}}, 
\code{\link{buildLib}} 
}
\examples{
### Load the inputData ###
data(exampleMetReport)
### Normalize ####
normalizedData <- normalizeByInternalStandard(
	exampleMetReport, 
	internalStandard = "Acetone", 
	save = FALSE
	)
### Show results ####
print(normalizedData)
}