% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStationMetabDo.R
\name{OneStationMetabDo}
\alias{OneStationMetabDo}
\title{R6 Class defining a one station metabolism model for DO}
\description{
A model for predicting dissolved oxygen concentrations in
   a stream using the one-station method.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{output}}{Model output data frame}

\item{\code{dailyGPP}}{Model parameter for daily gross primary production
based on C atoms fixed.
Units of micromoles per liter per day.}

\item{\code{ratioDoCfix}}{Ratio of DO molecules produced relative to carbon atoms fixed.}

\item{\code{dailyER}}{Model parameter for daily aerobic ecosystem respiration
based on carbon atoms respired.
Units of micromoles per liter per day.}

\item{\code{ratioDoCresp}}{Ratio of DO molecules consumed relative to carbon atoms respired.
Defaults to 1.}

\item{\code{k600}}{Model parameter for the gas exchange rate}

\item{\code{alphag}}{Model parameter for the turnover rate due to groundwater}

\item{\code{airPressure}}{Air pressure}

\item{\code{initialDO}}{Initial dissolved oxygen concentration}

\item{\code{time}}{Vector of numeric times (in days) for model predictions}

\item{\code{timePOSIX}}{Vector of times as POSIX object for model predictions}

\item{\code{temp}}{Vector of temperatures at the simulated times}

\item{\code{par}}{Vector of photosynthetically active radiation at the simulated times}

\item{\code{parTotal}}{Total PAR over the simulation period}

\item{\code{gwC}}{Concentration of oxygen in inflowing groundwater}

\item{\code{dt}}{Vector of time steps}

\item{\code{doSatCalculator}}{Function to use for do saturation calculations}

\item{\code{kSchmidtFunc}}{Function to use to adjust oxygen gas exchange for temperature}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OneStationMetabDo$new()}}
\item \href{#method-getDailyGppDo}{\code{OneStationMetabDo$getDailyGppDo()}}
\item \href{#method-getDailyErDo}{\code{OneStationMetabDo$getDailyErDo()}}
\item \href{#method-run}{\code{OneStationMetabDo$run()}}
\item \href{#method-plot}{\code{OneStationMetabDo$plot()}}
\item \href{#method-plotPAR}{\code{OneStationMetabDo$plotPAR()}}
\item \href{#method-clone}{\code{OneStationMetabDo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabDo$new(
  dailyGPP,
  ratioDoCfix = 1,
  dailyER,
  ratioDoCresp = -1,
  k600,
  alphag = NA,
  airPressure,
  initialDO,
  time,
  temp,
  par,
  parTotal = NA,
  stdAirPressure = 760,
  gwC = NA,
  doSatCalculator = DoSatCalculator$new(stdAirPressure = stdAirPressure),
  kSchmidtFunc = kSchmidtDO
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dailyGPP}}{Model parameter for daily gross primary production
based on C atoms fixed.
Units of micromoles per liter per day.}

\item{\code{ratioDoCfix}}{Ratio of DO molecules produced relative to carbon atoms fixed.
Defaults to 1.}

\item{\code{dailyER}}{Model parameter for daily aerobic ecosystem respiration
based on carbon atoms respired.
Units of micromoles per liter per day.}

\item{\code{ratioDoCresp}}{Ratio of DO molecules consumed relative to carbon atoms respired.
Defaults to -1.}

\item{\code{k600}}{Influence of atmospheric gas exchange on oxygen concentration as a first-order rate
depending on saturation deficit. Units are per day.}

\item{\code{alphag}}{Model parameter for the turnover rate due to groundwater in per day}

\item{\code{airPressure}}{Barometric pressure.
Units must be consistent with the units of the optional standard air pressure
argument.
Default value for standard air pressure results in units of mm Hg.}

\item{\code{initialDO}}{Initial DO concentration at the beginning of the evaluation period.
Units depend on the unit conversion factor for DO saturation.
Default value of the unit conversion factor results in
units of micomolarity.}

\item{\code{time}}{Vector of times associated with temperature and PAR data.
POSIXct vector or character vector in a standard text format
to be forced to POSIXct "YYYY-MM-DD HH:mm:ss".}

\item{\code{temp}}{Vector of water temperatures in degrees Celsius.
Must be same length as the time vector.}

\item{\code{par}}{Vector of photosynthetically active radiation values. 
Units are arbitrary, but must be consistent with total PAR if the
optional argument for total PAR is specified.
Must be the same length as the time vector.}

\item{\code{parTotal}}{total PAR, see units note in par argument description
(numerical vector with single value, defaults to NA which forces
calculation of total PAR from PAR data vector)}

\item{\code{stdAirPressure}}{The standard air pressure in the desired units. Defaults to 760 mm Hg.}

\item{\code{gwC}}{Concentration of oxygen in inflowing groundwater in micromolarity}

\item{\code{doSatCalculator}}{The DO saturation calculator used to estimate DO saturation concentrations
at a given temperature and air pressure. 
Defaults to a new instance of \code{\link{DoSatCalculator}} based on
the standard air pressure and saturation concentration unit conversion
arguments.}

\item{\code{kSchmidtFunc}}{Option to change the function used for calculating the gas exchange
rate from the gas exchange parameter normalized to a Schmidt number.
Defaults to \code{\link{kSchmidtDO}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDailyGppDo"></a>}}
\if{latex}{\out{\hypertarget{method-getDailyGppDo}{}}}
\subsection{Method \code{getDailyGppDo()}}{
Calculates the current daily GPP relative to DO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabDo$getDailyGppDo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Current daily GPP setting as a rate of change in DO concentration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDailyErDo"></a>}}
\if{latex}{\out{\hypertarget{method-getDailyErDo}{}}}
\subsection{Method \code{getDailyErDo()}}{
Calculates the current daily ER relative to DO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabDo$getDailyErDo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Current daily ER setting as a rate of change in DO concentration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Runs a one-station metabolism model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabDo$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with incremental and final results of the simulation,
   with columns:
   \itemize{
     \item time: POSIXct simulation time (tzone attribute will not be set)
     \item do: dissolved oxygen concentration in micromolality
     \item doSat: Saturated dissolved oxygen concentration in micromolality
     \item doProduction: increase in DO concentration in micromolality during the time step
     \item doConsumption: decrease in DO concentration in micromolality during the time step
     \item k: DO gas exchange rate in per day
     \item temp: water temperature in degrees Celsius
     \item dt: length of time step in days
   }
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the oxygen concentration, saturation oxygen concentration,
   and PAR on the same axes. Plots to the current active device.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabDo$plot(
  obsDO = numeric(),
  mar = c(2.5, 4.5, 1, 4),
  plotPAR = TRUE,
  colPAR = "red",
  labelPAR = "PAR",
  ylim = numeric(),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obsDO}}{Optional parameter to specify a vector of observations to plot
as points on the graph. The vector must be the same length as
the number of time steps in the model. Default value is an empty
vector, such that observations will only be plotted if a vector
is provided as an argument.}

\item{\code{mar}}{Optional parameter to change the inner margins of the plot. 
Default value is c(2.5, 4.5, 1, 4).}

\item{\code{plotPAR}}{Optional parameter to turn on/off plotting of PAR on the same axes.
Defaults to TRUE.}

\item{\code{colPAR}}{Optional parameter to set the color for the PAR plot.
Defaults to "red".}

\item{\code{labelPAR}}{Optional parameter for changing the PAR label.
Defaults to "PAR".}

\item{\code{ylim}}{Optional two-element vector to change the min/max scaling of the y axis
Defaults to empty vector, which results in the minimum and
maximum of oxygen data determining the axis scale}

\item{\code{...}}{Any additional parameters will be passed to the R "plot.default"
function used to set up the axes for the oxygen plots.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotPAR"></a>}}
\if{latex}{\out{\hypertarget{method-plotPAR}{}}}
\subsection{Method \code{plotPAR()}}{
Plots the PAR that was used for the model simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabDo$plotPAR(ylab = "PAR", col = "red", ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ylab}}{Optional parameter to change the y axis label.
Defaults to "PAR".}

\item{\code{col}}{Optional parameter to change the color of the plot.
Defaults to "red".}

\item{\code{...}}{Any additional parameters will be passed to the R "plot.default"
function used to set up the axes for the PAR plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabDo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
