% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStationMetabDoDic.R
\name{OneStationMetabDoDic}
\alias{OneStationMetabDoDic}
\title{R6 class defining a one station metabolism model for DO and DIC}
\description{
A model for predicting dissolved oxygen and carbon dioxide concentrations in
   a stream using the one station method.
}
\section{Super class}{
\code{\link[metab:OneStationMetabDo]{metab::OneStationMetabDo}} -> \code{OneStationMetabDoDic}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ratioDicCfix}}{Ratio of carbon atoms in DIC consumed relative to carbon atoms fixed.
Defaults to 1.}

\item{\code{dailyGPPDIC}}{Model parameter for daily gross primary production
based on C atoms in DIC consumed.
Units of micromoles per liter per day.}

\item{\code{ratioDicCresp}}{Ratio of carbon atoms in DIC produced relative to carbon atoms respired.
Defaults to 1.}

\item{\code{dailyERDIC}}{Model parameter for daily aerobic ecosystem respiration
based on C atoms in DIC consumed.
Units of micromoles per liter per day.}

\item{\code{initialDIC}}{initial total dissolved inorganic carbon concentration}

\item{\code{pCO2air}}{pCO2 in the atmosphere}

\item{\code{alkalinity}}{Total acid neutralizing capacity of the water}

\item{\code{carbonateEq}}{The carbonate equilibrium object to use for calculations}

\item{\code{kSchmidtFuncCO2}}{Function to use to adjust carbon dioxide gas exchange for temperature}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OneStationMetabDoDic$new()}}
\item \href{#method-run}{\code{OneStationMetabDoDic$run()}}
\item \href{#method-clone}{\code{OneStationMetabDoDic$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="OneStationMetabDo" data-id="plot">}\href{../../metab/html/OneStationMetabDo.html#method-plot}{\code{metab::OneStationMetabDo$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="OneStationMetabDo" data-id="plotPAR">}\href{../../metab/html/OneStationMetabDo.html#method-plotPAR}{\code{metab::OneStationMetabDo$plotPAR()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabDoDic$new(
  ...,
  ratioDicCfix = -1,
  ratioDicCresp = 1,
  initialDIC = NULL,
  initialpCO2 = NULL,
  pCO2air,
  alkalinity,
  kSchmidtFuncCO2 = kSchmidtCO2
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to constructor \code{OneStationMetabDoDic$new(...)} will be 
passed generically to the constructor for the superclass \code{OneStationMetabDo}. 
See documentation for the class \code{\link{OneStationMetabDo}} for a description
of these arguments.}

\item{\code{ratioDicCfix}}{Ratio of carbon atoms in DIC consumed relative to carbon atoms fixed.
Defaults to -1.}

\item{\code{ratioDicCresp}}{Ratio of carbon atoms in DIC produced relative to carbon atoms respired.
Defaults to 1.}

\item{\code{initialDIC}}{initial DIC concentration in micromoles per liter
(numerical vector, only first value will be used)
Either initialDIC or initialpCO2 must be specified, but not both.}

\item{\code{initialpCO2}}{initial pCO2 concentration in microatmospheres
(numerical vector, only first value will be used)
Either initialDIC or initialpCO2 must be specified, but not both.}

\item{\code{pCO2air}}{partial pressure of CO2 in the air in microatmospheres
(numerical vector)}

\item{\code{alkalinity}}{alkalinity of stream water
(numerical vector)}

\item{\code{kSchmidtFuncCO2}}{Option to change the function used for calculating the carbon dioxide gas exchange
rate from the gas exchange parameter normalized to a Schmidt number.
Defaults to \code{\link{kSchmidtCO2}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Runs the model predicting the change in DO and DIC concentrations over
   time at a location along a stream.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabDoDic$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with incremental and final results of the simulation,
   with columns \cr
   \itemize{
     \item All columns generated by \code{\link{OneStationMetabDo}} run method
     \item co2Production: increase in co2 concentration during the time step in
           micromoles per liter
     \item co2Consumption: decrease in co2 concentration during the time step in
           micromoles per liter
     \item kH: Henry's constant during the time step in micromoles per liter per 
           microatmospheres
     \item co2Sat: saturation concentration for co2 in micromoles per liter
     \item fGas: change in co2 concentration due to air-water exchange in
           micromoles per liter
     \item pH: minus the log of the hydrogen ion concentration
     \item pCO2: partial pressure of co2 in water in microatmosphere
     \item dic: DIC concentration in water in moles per liter 
   }
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabDoDic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
