% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStationMetabExtractor.R
\name{OneStationMetabExtractor}
\alias{OneStationMetabExtractor}
\title{R6 class for extracting results from one station metabolism models}
\description{
Extracts parameters used from metabolism models
}
\section{Super class}{
\code{\link[disco:SignalSummarizer]{disco::SignalSummarizer}} -> \code{OneStationMetabExtractor}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{table}}{Table of summary data}

\item{\code{index}}{Index for tracking the row in the table}

\item{\code{resultFile}}{The name of the file with analysis results}

\item{\code{summaryFile}}{The name of the file to create with the summary}

\item{\code{path}}{Table of summary data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OneStationMetabExtractor$new()}}
\item \href{#method-open}{\code{OneStationMetabExtractor$open()}}
\item \href{#method-summarize}{\code{OneStationMetabExtractor$summarize()}}
\item \href{#method-close}{\code{OneStationMetabExtractor$close()}}
\item \href{#method-clone}{\code{OneStationMetabExtractor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabExtractor$new(
  rows,
  resultFile = "results",
  summaryFile = "summary"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows}}{number of rows in the table of extracted parameter values}

\item{\code{resultFile}}{name of the results files to read}

\item{\code{summaryFile}}{name of the summary file to create}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open"></a>}}
\if{latex}{\out{\hypertarget{method-open}{}}}
\subsection{Method \code{open()}}{
Prepare to write the summary file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabExtractor$open(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path where the summary file will be written}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
Collects the data for the summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabExtractor$summarize(signal, outputPath, label, timeBounds)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal}}{The signal being summarized}

\item{\code{outputPath}}{The path where analysis output is located}

\item{\code{label}}{A label for the summary}

\item{\code{timeBounds}}{The temporal bounds on the summary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
Closes the summary output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabExtractor$close()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabExtractor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
