% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStationMetabMLE.R
\name{OneStationMetabMLE}
\alias{OneStationMetabMLE}
\title{R6 class defining an MLE inference using a one station metabolism model}
\description{
Processes a signal of DO and pCO2 data 
  (with supporting PAR and temperature information)
  for inferring whole-stream metabolism using the one-station
  method.
}
\section{Super class}{
\code{\link[disco:SignalDerivation]{disco::SignalDerivation}} -> \code{OneStationMetabMLE}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{initParams}}{The intial parameter values to use for the MLE algorithm}

\item{\code{parameterTranslatorGenerator}}{The translator used to convert the parameters provided into
model input}

\item{\code{isEstimated}}{A vector of logical values indicating which parameters
to include in the estimation.}

\item{\code{staticAirPressure}}{The air pressure.}

\item{\code{sdLikelihood}}{The standard deviations used to calculate likelihood}

\item{\code{useDO}}{Set to TRUE to use DO in the inference.}

\item{\code{usepCO2}}{Set to TRUE to use pCO2 in the inference.}

\item{\code{outputFile}}{The name of the output file.}

\item{\code{staticCO2Air}}{The pCO2 in the air.}

\item{\code{staticAlkalinity}}{The alkalinity.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OneStationMetabMLE$new()}}
\item \href{#method-derive}{\code{OneStationMetabMLE$derive()}}
\item \href{#method-clone}{\code{OneStationMetabMLE$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabMLE$new(
  ...,
  initParams,
  parameterTranslatorGenerator = ParameterTranslatorMetab,
  isEstimated = c(TRUE, TRUE, TRUE),
  sdLikelihood,
  useDO = TRUE,
  usepCO2 = FALSE,
  staticAirPressure = NULL,
  outputFile = "results",
  staticCO2Air = NULL,
  staticAlkalinity = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments needed by the constructor of the super class}

\item{\code{initParams}}{The intial parameter values to use for the MLE algorithm}

\item{\code{parameterTranslatorGenerator}}{The R6 class generator for the translator used to convert the parameters 
provided into model input.
Default value is the R6 class generator ParameterTranslatorMetab.}

\item{\code{isEstimated}}{A vector of logical values indicating which parameters
to include in the estimation.}

\item{\code{sdLikelihood}}{The standard deviations used to calculate likelihood}

\item{\code{useDO}}{Set to TRUE to use DO in the inference.
Defaults to TRUE.}

\item{\code{usepCO2}}{Set to TRUE to use pCO2 in the inference.
Defaults to FALSE.}

\item{\code{staticAirPressure}}{The air pressure.
Defaults to NULL, which will cause use of air pressure
from the signal.}

\item{\code{outputFile}}{The name of the output file.
Defaults to "results".}

\item{\code{staticCO2Air}}{The pCO2 in the air.
Defaults to NULL, which will cause use of pCO2
from the signal.}

\item{\code{staticAlkalinity}}{The alkalinity.
Defaults to NULL, which will cause use of alkalinity
from the signal.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-derive"></a>}}
\if{latex}{\out{\hypertarget{method-derive}{}}}
\subsection{Method \code{derive()}}{
Uses a one-station model to infer whole-stream metabolism and
  gas exchange parameters from a single signal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabMLE$derive(signal = NULL, prevResults = NULL, path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal}}{The signal on which the MLE inference is based}

\item{\code{prevResults}}{The results from the previous inference}

\item{\code{path}}{The path to where the results should be written}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The two-element list containint the results of the MLE.
  \itemize{
    \item objFunc: The objective function used for the MLE
    \item optimr: The results from optim performing the MLE
  }
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabMLE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
