% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStationMetabPlotter.R
\name{OneStationMetabPlotter}
\alias{OneStationMetabPlotter}
\title{R6 class defining a plotter for a one station metabolism DO model}
\description{
Plots a signal of DO and pCO2 from a one station model of metabolism
}
\section{Super classes}{
\code{\link[disco:SignalSummarizer]{disco::SignalSummarizer}} -> \code{\link[disco:SignalPlotter]{disco::SignalPlotter}} -> \code{OneStationMetabPlotter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{timeTicks}}{Number of ticks on the time axis}

\item{\code{resultFile}}{Base file name for the results.}

\item{\code{results}}{The results object to be plotted}

\item{\code{plotDIC}}{Flag to determine if DIC should be plotted rather
than pCO2.}

\item{\code{subTitle}}{Subtitle for headings on plots, will appear over right
panel on legend.}

\item{\code{format}}{Data frame aggregating the formats for all the plots}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{OneStationMetabPlotter$new()}}
\item \href{#method-open}{\code{OneStationMetabPlotter$open()}}
\item \href{#method-summarize}{\code{OneStationMetabPlotter$summarize()}}
\item \href{#method-clone}{\code{OneStationMetabPlotter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="disco" data-topic="SignalPlotter" data-id="close">}\href{../../disco/html/SignalPlotter.html#method-close}{\code{disco::SignalPlotter$close()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabPlotter$new(
  ...,
  timeTicks,
  resultFile = "results",
  plotDIC = FALSE,
  subTitle = "",
  format.doObs = list(name = "[DO] observed", lty = NA, pch = 1, col = "black", lwd =
    1),
  format.pCO2Obs = list(name = "pCO2 or DIC observed", lty = NA, pch = 1, col =
    "green3", lwd = 1),
  format.dicObs = list(name = "DIC Obs", lty = NA, pch = 1, col = "green3", lwd = 1),
  format.doPred = list(name = "[DO] modeled", lty = "solid", pch = NA, col = "black",
    lwd = 2),
  format.pCO2Pred = list(name = "pCO2 or DIC modeled", lty = "solid", pch = NA, col =
    "green3", lwd = 2),
  format.dicPred = list(name = "DIC Pred", lty = "solid", pch = NA, col = "green3", lwd
    = 2),
  format.doSat = list(name = "Saturated [DO]", lty = "dotted", pch = NA, col = "black",
    lwd = 2),
  format.par = list(name = "PAR", lty = "dashed", pch = NA, col = "red", lwd = 2),
  format.temp = list(name = "Temperature", lty = NA, pch = 1, col = "black", lwd = 1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to the constructor of the super class}

\item{\code{timeTicks}}{Number of ticks on the time axis}

\item{\code{resultFile}}{Base file name for the results.
Defaults to "results".}

\item{\code{plotDIC}}{Flag to determine if DIC should be plotted rather
than pCO2.
Defaults to FALSE.}

\item{\code{subTitle}}{Subtitle for headings on plots, will appear over right
panel on legend.
Defaults to an empty string (blank).}

\item{\code{format.doObs}}{Format for DO observations.
See defaults in usage.}

\item{\code{format.pCO2Obs}}{Format for pCO2 observations.
See defaults in usage.}

\item{\code{format.dicObs}}{Format for DIC observations.
See defaults in usage.}

\item{\code{format.doPred}}{Format for DO predictions
See defaults in usage.}

\item{\code{format.pCO2Pred}}{Format for pCO2 predictions
See defaults in usage.}

\item{\code{format.dicPred}}{Format for DIC predictions
See defaults in usage.}

\item{\code{format.doSat}}{Format for saturation DO concentrations
See defaults in usage.}

\item{\code{format.par}}{Format for PAR
See defaults in usage.}

\item{\code{format.temp}}{Format for temperature
See defaults in usage.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open"></a>}}
\if{latex}{\out{\hypertarget{method-open}{}}}
\subsection{Method \code{open()}}{
Opens the pdf for writing the plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabPlotter$open(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path where the pdf of figures should be written}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
Summarize the data in a two panel visualization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabPlotter$summarize(
  signal = NULL,
  outputPath = NULL,
  label,
  timeBounds
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal}}{The signal object that contains observations}

\item{\code{outputPath}}{A path to results of a metabolism analysis.
Not necessary if summarizer does not need output.}

\item{\code{label}}{A label for the summary}

\item{\code{timeBounds}}{The temporal bounds on the summary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OneStationMetabPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
