% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferno.R
\name{ParameterTranslatorMetab}
\alias{ParameterTranslatorMetab}
\title{R6 class defining a parameter translator for metabolism models}
\description{
A parameter translator allowing GPP, ER, and k600 to be used
   as proposed parameters for an objective function using a
   stream metabolism model.
}
\section{Super class}{
\code{\link[inferno:ParameterTranslator]{inferno::ParameterTranslator}} -> \code{ParameterTranslatorMetab}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isEstimated}}{A vector of logical values indicating which parameters
to include in the estimation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParameterTranslatorMetab$new()}}
\item \href{#method-translate}{\code{ParameterTranslatorMetab$translate()}}
\item \href{#method-clone}{\code{ParameterTranslatorMetab$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="ParameterTranslator" data-id="setModel">}\href{../../inferno/html/ParameterTranslator.html#method-setModel}{\code{inferno::ParameterTranslator$setModel()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterTranslatorMetab$new(..., isEstimated = c(TRUE, TRUE, TRUE))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments to pass to constructor of super class}

\item{\code{isEstimated}}{A vector of logical values indicating which parameters
to include in the estimation.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-translate"></a>}}
\if{latex}{\out{\hypertarget{method-translate}{}}}
\subsection{Method \code{translate()}}{
Used to translate 3 parameters in the order of
  gross primary production, ecosystem respiration,
  and gas exchange at a Schmidt number of 600
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterTranslatorMetab$translate(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Vector of values to use for GPP, ER, and k600 parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParameterTranslatorMetab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
