% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inferno.R
\name{PredictionExtractorMetabDoDic}
\alias{PredictionExtractorMetabDoDic}
\title{R6 class defining a prediction extractor for DO and pCO2 from metabolism models}
\description{
A prediction extractor allowing dissolved oxygen and carbo
   dioxide outputs from a stream metabolism model to be used as 
   predictions in an objective function.
}
\section{Super class}{
\code{\link[inferno:PredictionExtractor]{inferno::PredictionExtractor}} -> \code{PredictionExtractorMetabDoDic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-extract}{\code{PredictionExtractorMetabDoDic$extract()}}
\item \href{#method-clone}{\code{PredictionExtractorMetabDoDic$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="PredictionExtractor" data-id="initialize">}\href{../../inferno/html/PredictionExtractor.html#method-initialize}{\code{inferno::PredictionExtractor$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract"></a>}}
\if{latex}{\out{\hypertarget{method-extract}{}}}
\subsection{Method \code{extract()}}{
Used to extract dissolved oxygen and dissolved carbon dioxide
  predictions from a metabolism model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExtractorMetabDoDic$extract()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with two columns for DO and pCO2 data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExtractorMetabDoDic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
