% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsUtilities.R
\name{SolarRadiation}
\alias{SolarRadiation}
\title{R6 class defining incoming solar radiation}
\description{
Provides the ability to make solar radiation calculations
  for a given location on Earth.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{latitudeAngle}}{Angle of the latitude}

\item{\code{longitudeAngle}}{Angle of the longitude}

\item{\code{differenceFromGMT}}{Number of hours different from UTC}

\item{\code{adjustDST}}{Number of hours to adjust during daylight savings}

\item{\code{solarConstantFlux}}{The solar constant for flux of radiation (for Earth)}

\item{\code{earthAngularVelocity}}{The angular velocity of Earth's rotation on its axis}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SolarRadiation$new()}}
\item \href{#method-getExtraterrestrialInsolation}{\code{SolarRadiation$getExtraterrestrialInsolation()}}
\item \href{#method-clone}{\code{SolarRadiation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarRadiation$new(
  latitude,
  longitude,
  differenceFromGMT,
  adjustDST = NA,
  solarConstantFlux = 1364,
  earthAngularVelocity = 0.2618
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{latitude}}{Latitude on earth in decimal degrees}

\item{\code{longitude}}{Longitude on earth in decimal degrees}

\item{\code{differenceFromGMT}}{Number of hours time zone is different from GMT}

\item{\code{adjustDST}}{Number of hours to adjust the difference from GMT during 
daylight savings time.
Default value is NA, which disables daylight savings adjustments.}

\item{\code{solarConstantFlux}}{Base flux of energy from the sun.
Default value is 1364 Watts per square meter}

\item{\code{earthAngularVelocity}}{Angular velocity of earth's rotation in radians per hour.
Default value is 0.2618.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getExtraterrestrialInsolation"></a>}}
\if{latex}{\out{\hypertarget{method-getExtraterrestrialInsolation}{}}}
\subsection{Method \code{getExtraterrestrialInsolation()}}{
Calculates the incoming extraterrestrial solar radiation at
  the times provided
  (i.e. before any influence by atmosphere) for the location
  on earth represented by the SolarRadiation object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarRadiation$getExtraterrestrialInsolation(
  time,
  timeCT = as.POSIXct(time),
  timeLT = as.POSIXlt(timeCT),
  solarNoonCorrectionTime = 2 * pi * (timeLT$yday - 81)/365,
  solarNoonCorrectionEcc = (9.87 * sin(2 * solarNoonCorrectionTime) - 7.53 *
    cos(solarNoonCorrectionTime) - 1.5 * sin(solarNoonCorrectionTime))/60,
  dayAngle = 2 * pi * (timeLT$yday/365),
  eccCoefficient = 1.00011 + 0.034221 * cos(dayAngle) + 0.00128 * sin(dayAngle) +
    0.000719 * cos(2 * dayAngle) + 7.7e-05 * sin(2 * dayAngle),
  declinationAngle = 0.006918 - 0.399912 * cos(dayAngle) + 0.070257 * sin(dayAngle) -
    0.006758 * cos(2 * dayAngle) + 0.000907 * sin(2 * dayAngle) - 0.002697 * cos(3 *
    dayAngle) + 0.00148 * sin(3 * dayAngle)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The time(s) at which solar radiation should be calculated.
Must be coercable into a POSIX type time.}

\item{\code{timeCT}}{The time in POSIXct type.
Default value is time coerced into POSIXct class.}

\item{\code{timeLT}}{The time in POSXlt type.
Default value is timeCT coerced into POSIXlt class.}

\item{\code{solarNoonCorrectionTime}}{The time basis for calculating solar noon corrections.
Defaults to a conversion to radians from a version of the
day angle.}

\item{\code{solarNoonCorrectionEcc}}{The correction for eccentricities in Earth's orbit.
Defaults to an empirical approximation from solarNoonCorrectionTime}

\item{\code{dayAngle}}{The angle representing the day of the year, if the year is represented
by 2 * pi radians
Default value is calculated based on the day of year form timeLT.}

\item{\code{eccCoefficient}}{The coefficient of correction for radiation based on eccentricities in
Earth's distance from the sun during its orbit. 
Default value is an empirical calculation from the day angle.}

\item{\code{declinationAngle}}{The declination angle of the sun relative to Earth's equator.
Default value is an empirical calculation from the day angle.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The solar radiation flux at the top of Earth's atmosphere
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SolarRadiation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
