% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStationMetabDo.R
\name{TwoStationMetabDo}
\alias{TwoStationMetabDo}
\title{R6 class defining a two station metabolism model for DO}
\description{
A model for predicting changes in dissolved oxygen concentrations
   over a reach of a stream using the two-station method.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{output}}{Data frame with output from the model}

\item{\code{dailyGPP}}{Model parameter for daily gross primary production
based on C atoms fixed.
Units of micromoles per liter per day.}

\item{\code{ratioDoCfix}}{Ratio of DO molecules produced relative to carbon atoms fixed.}

\item{\code{dailyER}}{Model parameter for daily aerobic ecosystem respiration
based on carbon atoms respired.
Units of micromoles per liter per day.}

\item{\code{ratioDoCresp}}{Ratio of DO molecules consumed relative to carbon atoms respired.
Defaults to 1.}

\item{\code{k600}}{Model parameter for the gas exchange rate}

\item{\code{alphag}}{Model parameter for the turnover rate due to groundwater}

\item{\code{airPressure}}{Air pressure}

\item{\code{par}}{Vector of photosynthetically active radiation at the simulated times}

\item{\code{parTotal}}{Total PAR over the simulation period}

\item{\code{upstreamTime}}{Numeric time in days a parcel of water is passing upstream station}

\item{\code{upstreamTimePOSIX}}{POSIX time a parcel of water is passing upstream station}

\item{\code{upstreamTemp}}{Temperature of a parcel of water when passing upstream station}

\item{\code{upstreamPAR}}{Photosynthetically active radiation on a parcel of water when 
passing upstream station}

\item{\code{upstreamDO}}{DO concentration of a parcel of water when passing upstream station}

\item{\code{upstreamDOSat}}{The saturation concentration of DO of a parcel of water when
passing upstream station}

\item{\code{downstreamTime}}{Time a parcel of water is passing downstream station}

\item{\code{downstreamTimePOSIX}}{POSIX time a parcel of water is passing upstream station}

\item{\code{downstreamTemp}}{Temperature of a parcel of water when passing downstream station}

\item{\code{downstreamPAR}}{Photosynthetically active radiation on a parcel of water when 
passing downstream station}

\item{\code{downstreamDOSat}}{The saturation concentration of DO of a parcel of water when
passing downstream station}

\item{\code{gwC}}{Concentration of oxygen in inflowing groundwater}

\item{\code{doSatCalculator}}{Function to use for do saturation calculations}

\item{\code{kSchmidtFunc}}{Function to use to adjust gas exchange for temperature}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TwoStationMetabDo$new()}}
\item \href{#method-getDailyGppDo}{\code{TwoStationMetabDo$getDailyGppDo()}}
\item \href{#method-getDailyErDo}{\code{TwoStationMetabDo$getDailyErDo()}}
\item \href{#method-run}{\code{TwoStationMetabDo$run()}}
\item \href{#method-plot}{\code{TwoStationMetabDo$plot()}}
\item \href{#method-plotDO}{\code{TwoStationMetabDo$plotDO()}}
\item \href{#method-plotTemp}{\code{TwoStationMetabDo$plotTemp()}}
\item \href{#method-clone}{\code{TwoStationMetabDo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDo$new(
  dailyGPP,
  ratioDoCfix = 1,
  dailyER,
  ratioDoCresp = -1,
  k600,
  alphag = NA,
  airPressure,
  upstreamTime,
  upstreamTemp,
  upstreamPAR = NA,
  upstreamDO,
  downstreamTime,
  downstreamTemp,
  downstreamPAR = NA,
  par = 0.5 * (upstreamPAR + downstreamPAR),
  parTotal = NA,
  stdAirPressure = 760,
  gwC = NA,
  doSatCalculator = DoSatCalculator$new(stdAirPressure = stdAirPressure),
  kSchmidtFunc = kSchmidtDO
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dailyGPP}}{Daily average influence of gross primary production on oxygen concentration.
Units depend on the unit conversion factor for DO saturation.
Default value of the unit conversion factor results in
units of micomolarity per day.}

\item{\code{ratioDoCfix}}{Ratio of DO molecules produced relative to carbon atoms fixed.
Defaults to 1.}

\item{\code{dailyER}}{Daily average influence of ecosystem respiration on oxygen.
Units depend on the unit conversion factor for DO saturation.
Default value of the unit conversion factor results in
units of micomolarity per day.}

\item{\code{ratioDoCresp}}{Ratio of DO molecules consumed relative to carbon atoms respired.
Defaults to -1.}

\item{\code{k600}}{Influence of atmospheric gas exchange on oxygen concentration as a first-order rate
depending on saturation deficit. Units are per day.}

\item{\code{alphag}}{Model parameter for the turnover rate due to groundwater in per day}

\item{\code{airPressure}}{Barometric pressure.
Units must be consistent with the units of the optional standard air pressure
argument.
Default value for standard air pressure results in units of mm Hg.}

\item{\code{upstreamTime}}{Time vector for data at upstream location.
(POSIXct vector or character vector in a standard text format
to be forced to POSIXct "YYYY-MM-DD HH:mm:ss")}

\item{\code{upstreamTemp}}{Vector of temperatures in degrees C at upstream location.}

\item{\code{upstreamPAR}}{Optional vector of values for PAR at upstream station}

\item{\code{upstreamDO}}{Vector of DO concentrations at upstream location.}

\item{\code{downstreamTime}}{Time vector for data at downstream location.
In relation to upstream times, the downstream times should reflect
the travel times of water over the reach.
(POSIXct vector or character vector in a standard text format
to be forced to POSIXct "YYYY-MM-DD HH:mm:ss")}

\item{\code{downstreamTemp}}{Vector of temperatures in degrees C at the downstream location.}

\item{\code{downstreamPAR}}{Optional vector of values for PAR at downstream station}

\item{\code{par}}{Photosynthetically active radiation. 
Units are arbitrary, but must be consistent with total PAR total if the
optional argument for total PAR is specified.
Defaults to the average of upstream and downstream PAR arguments}

\item{\code{parTotal}}{Total PAR over the evaluation period. Units must be the same as the PAR
data argument.
Default is the integration of the PAR data provided over the evaluation period.}

\item{\code{stdAirPressure}}{The standard air pressure in the desired units. 
Defaults to 760 mm Hg.}

\item{\code{gwC}}{Concentration of oxygen in inflowing groundwater in micromolarity}

\item{\code{doSatCalculator}}{The DO saturation calculator used to estimate DO saturation concentrations
at a given temperature and air pressure. 
Defaults to a new instance of \code{\link{DoSatCalculator}} based on
the standard air pressure and saturation concentration unit conversion
arguments.}

\item{\code{kSchmidtFunc}}{Option to change the function used for calculating the gas exchange
rate from the gas exchange parameter normalized to a Schmidt number.
Defaults to \code{\link{kSchmidtDO}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDailyGppDo"></a>}}
\if{latex}{\out{\hypertarget{method-getDailyGppDo}{}}}
\subsection{Method \code{getDailyGppDo()}}{
Calculates the current daily GPP relative to DO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDo$getDailyGppDo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Current daily GPP setting as a rate of change in DO concentration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDailyErDo"></a>}}
\if{latex}{\out{\hypertarget{method-getDailyErDo}{}}}
\subsection{Method \code{getDailyErDo()}}{
Calculates the current daily ER relative to DO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDo$getDailyErDo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Current daily ER setting as a rate of change in DO concentration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Runs a two-station metabolism model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDo$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with incremental and final results of the simulation,
   with columns:
   \itemize{
     \item time: POSIXct simulation time
     \item residenceTime: length of time between upstream and downstream times
     \item do: dissolved oxygen concentration
     \item doProduction: increase in DO concentration in travel over the reach
     \item doConsumption: decrease in DO concentration in travel over the reach
     \item doEquilibration: two-station term for gas exchange in travel over the reach
     \item k: DO gas exchange applied in travel over the reach
   }
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the results of a two station model simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDo$plot(
  obs.DO = numeric(),
  ylim.DO = numeric(),
  mar.DO = c(2.5, 4.5, 2, 4),
  ylim.temp = numeric(),
  mar.temp = c(2.5, 4.5, 2, 4),
  mfrow = c(1, 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obs.DO}}{vector of observed DO concentrations.
Defaults to an empty vector which disables plotting of observations.}

\item{\code{ylim.DO}}{The limits of the scale on the y axis for DO concentrations.
Defaults to an empty vector which calculates the min and max from
the simulated and observed values.}

\item{\code{mar.DO}}{The margins on the DO plot.
Defaults to c(2.5, 4.5, 2, 4).}

\item{\code{ylim.temp}}{The limits of the scale on the y axis for temperature.
Defaults to an empty vector which calculates the min and max from
the temperature data.}

\item{\code{mar.temp}}{The margins on the temperature plot.
Defaults to c(2.5, 4.5, 2, 4).}

\item{\code{mfrow}}{The structure for multi-panel plots.
Defaults to c(1, 2), which is one row and two columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotDO"></a>}}
\if{latex}{\out{\hypertarget{method-plotDO}{}}}
\subsection{Method \code{plotDO()}}{
Plot the simulated DO concentrations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDo$plotDO(
  obs = numeric(),
  ylim = numeric(),
  mar = c(2.5, 4.5, 1, 4),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obs}}{vector of observed DO concentrations.
Defaults to an empty vector which disables plotting of observations.}

\item{\code{ylim}}{The limits of the scale on the y axis for DO concentrations.
Defaults to an empty vector which calculates the min and max from
the simulated and observed values.}

\item{\code{mar}}{The margins of the DO plot.
Defaults to c(2.5, 4.5, 1, 4).}

\item{\code{...}}{Other abstract arguments are passed to the plot.default function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotTemp"></a>}}
\if{latex}{\out{\hypertarget{method-plotTemp}{}}}
\subsection{Method \code{plotTemp()}}{
Plot the temperature.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDo$plotTemp(ylim = numeric(), mar = c(2.5, 4.5, 1, 4), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ylim}}{The limits of the scale on the y axis for temperature.
Defaults to an empty vector which calculates the min and max from
the temperature data.}

\item{\code{mar}}{The margins of the temperature plot.
Defaults to c(2.5, 4.5, 1, 4).}

\item{\code{...}}{Abstract arguments to be passed to the plot.default function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
