% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStationMetabDoDic.R
\name{TwoStationMetabDoDic}
\alias{TwoStationMetabDoDic}
\title{R6 class defining a two station metabolism model for DO and DIC}
\description{
A model for predicting changes in dissolved oxygen concentrations
   and dissolved inorganic carbon concentrations
   over a reach of a stream using the two-station method.
}
\section{Super class}{
\code{\link[metab:TwoStationMetabDo]{metab::TwoStationMetabDo}} -> \code{TwoStationMetabDoDic}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ratioDicCfix}}{Ratio of carbon atoms in DIC consumed relative to carbon atoms fixed.
Defaults to 1.}

\item{\code{ratioDicCresp}}{Ratio of carbon atoms in DIC produced relative to carbon atoms respired.
Defaults to 1.}

\item{\code{upstreamDIC}}{Upstream DIC concentration}

\item{\code{upstreampCO2}}{Upstream partial pressure of carbon dioxide}

\item{\code{upstreampH}}{Upstream pH}

\item{\code{upstreamkH}}{Upstream Henry's constant}

\item{\code{upstreamCO2Sat}}{Upstream saturation concentration for CO2}

\item{\code{upstreamDICSat}}{Upstream DIC at saturation concentration for CO2}

\item{\code{downstreamkH}}{Downstream Henry's constant}

\item{\code{downstreamCO2Sat}}{Downstream saturation concentration for CO2}

\item{\code{downstreamDICSat}}{Downstream DIC at saturation concentration for CO2}

\item{\code{pCO2air}}{Partial pressure of CO2 in the air}

\item{\code{alkalinity}}{Acid neutralizing capacity of the water}

\item{\code{kSchmidtFuncCO2}}{Function to use to adjust carbon dioxide gas exchange for temperature}

\item{\code{maxDIC}}{Maximum DIC to allow in optimization}

\item{\code{carbonateEq}}{The carbonate equilibrium object to use for calculations}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TwoStationMetabDoDic$new()}}
\item \href{#method-getDailyGppDic}{\code{TwoStationMetabDoDic$getDailyGppDic()}}
\item \href{#method-getDailyErDic}{\code{TwoStationMetabDoDic$getDailyErDic()}}
\item \href{#method-run}{\code{TwoStationMetabDoDic$run()}}
\item \href{#method-plot}{\code{TwoStationMetabDoDic$plot()}}
\item \href{#method-plotpCO2}{\code{TwoStationMetabDoDic$plotpCO2()}}
\item \href{#method-clone}{\code{TwoStationMetabDoDic$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="TwoStationMetabDo" data-id="getDailyErDo">}\href{../../metab/html/TwoStationMetabDo.html#method-getDailyErDo}{\code{metab::TwoStationMetabDo$getDailyErDo()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="TwoStationMetabDo" data-id="getDailyGppDo">}\href{../../metab/html/TwoStationMetabDo.html#method-getDailyGppDo}{\code{metab::TwoStationMetabDo$getDailyGppDo()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="TwoStationMetabDo" data-id="plotDO">}\href{../../metab/html/TwoStationMetabDo.html#method-plotDO}{\code{metab::TwoStationMetabDo$plotDO()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="metab" data-topic="TwoStationMetabDo" data-id="plotTemp">}\href{../../metab/html/TwoStationMetabDo.html#method-plotTemp}{\code{metab::TwoStationMetabDo$plotTemp()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDoDic$new(
  ...,
  ratioDicCfix = -1,
  ratioDicCresp = 1,
  upstreamDIC = NULL,
  upstreampCO2 = NULL,
  pCO2air,
  alkalinity,
  kSchmidtFuncCO2 = kSchmidtCO2,
  maxDIC = 1e+06
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to constructor \code{TwoStationMetabDoDic$new(...)} will be 
passed generically to the constructor for the superclass \code{TwoStationMetabDo}. 
See documentation for the constructor of class \code{\link{TwoStationMetabDo}} for a description
of these arguments.}

\item{\code{ratioDicCfix}}{Ratio of carbon atoms in DIC consumed relative to carbon atoms fixed.
Defaults to -1.}

\item{\code{ratioDicCresp}}{Ratio of carbon atoms in DIC produced relative to carbon atoms respired.
Defaults to 1.}

\item{\code{upstreamDIC}}{upstream DIC concentration in micromoles per liter
(numerical vector the same length as the upstream time vector)
Either upstreamDIC or upstreampCO2 must be specified, but not both.}

\item{\code{upstreampCO2}}{upstream pCO2 concentration in microatmospheres
(numerical vector the same length as the upstream time vector)
Either upstreamDIC or upstreampCO2 must be specified, but not both.}

\item{\code{pCO2air}}{partial pressure of CO2 in the air in microatmospheres
(numerical vector)}

\item{\code{alkalinity}}{alkalinity of stream water
(numerical vector)}

\item{\code{kSchmidtFuncCO2}}{Option to change the function used for calculating the carbon dioxide gas exchange
rate from the gas exchange parameter normalized to a Schmidt number.
Defaults to \code{\link{kSchmidtCO2}}}

\item{\code{maxDIC}}{Maximum DIC to allow in optimization.
Defaults to 1e6.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDailyGppDic"></a>}}
\if{latex}{\out{\hypertarget{method-getDailyGppDic}{}}}
\subsection{Method \code{getDailyGppDic()}}{
Calculates the current daily GPP relative to DIC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDoDic$getDailyGppDic()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Current daily GPP setting as a rate of change in DIC concentration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDailyErDic"></a>}}
\if{latex}{\out{\hypertarget{method-getDailyErDic}{}}}
\subsection{Method \code{getDailyErDic()}}{
Calculates the current daily ER relative to DIC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDoDic$getDailyErDic()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Current daily ER setting as a rate of change in DIC concentration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Runs a two-station metabolism model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDoDic$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with incremental and final results of the simulation,
   with columns:
   \itemize{
     \item All columns generated by the run method of \code{\link{TwoStationMetabDo}}
     \item co2Production: concentration change in co2 due to respiration
     \item co2Consumption: concentration change in co2 due to GPP
     \item co2Equilibration: two-station co2 equilibration term from gas exchange calculation
     \item pH: the downstream pH (minus the log base 10 of hydrogen ion concentration)
     \item pCO2: the downstream partial pressure of carbon dioxide
     \item dic: the downstream total dissolved inorganic carbon concentration
   }
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the results of the model simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDoDic$plot(
  obs.DO = numeric(),
  ylim.DO = numeric(),
  mar.DO = c(2.5, 4.5, 2, 4),
  obs.pCO2 = numeric(),
  ylim.pCO2 = numeric(),
  mar.pCO2 = mar.DO,
  mfrow = c(1, 2)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obs.DO}}{vector of observed DO concentrations.
Defaults to an empty vector which disables plotting of observations.}

\item{\code{ylim.DO}}{The limits of the scale on the y axis for DO concentrations.
Defaults to an empty vector which calculates the min and max from
the simulated and observed values.}

\item{\code{mar.DO}}{The margins on the DO plot.
Defaults to c(2.5, 4.5, 2, 4).}

\item{\code{obs.pCO2}}{vector of observed CO2 partial pressures.
Defaults to an empty vector which disables plotting of observations.}

\item{\code{ylim.pCO2}}{The limits of the scale on the y axis for CO2 partial pressure.
Defaults to an empty vector which calculates the min and max from
the simulated and observed values.}

\item{\code{mar.pCO2}}{The margins on the CO2 partial pressure plot.
Defaults to the value of the mar.DO argument.}

\item{\code{mfrow}}{The structure for multi-panel plots.
Defaults to c(1, 2), which is one row and two columns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotpCO2"></a>}}
\if{latex}{\out{\hypertarget{method-plotpCO2}{}}}
\subsection{Method \code{plotpCO2()}}{
Plot the partial pressure of CO2 simulation from the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDoDic$plotpCO2(
  obs = numeric(),
  ylim = numeric(),
  mar = c(2.5, 4.5, 1, 4),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obs}}{vector of observed CO2 partial pressures.
Defaults to an empty vector which disables plotting of observations.}

\item{\code{ylim}}{The limits of the scale on the y axis for CO2 partial pressures.
Defaults to an empty vector which calculates the min and max from
the simulated and observed values.}

\item{\code{mar}}{The margins of the CO2 partial pressure plot.
Defaults to c(2.5, 4.5, 1, 4).}

\item{\code{...}}{Other abstract arguments are passed to the plot.default function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabDoDic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
