% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStationMetabMLE.R
\name{TwoStationMetabMLE}
\alias{TwoStationMetabMLE}
\title{R6 class defining an MLE inference using a two station metabolism model}
\description{
Basic MLE inference for two station stream metabolism
}
\section{Super class}{
\code{\link[disco:TransferFunctionDerivation]{disco::TransferFunctionDerivation}} -> \code{TwoStationMetabMLE}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{initParams}}{The intial parameter values to use for the MLE algorithm}

\item{\code{parameterTranslatorGenerator}}{The translator used to convert the parameters provided into
model input}

\item{\code{isEstimated}}{A vector of logical values indicating which parameters
to include in the estimation.}

\item{\code{staticAirPressure}}{The air pressure.}

\item{\code{sdLikelihood}}{The standard deviations used to calculate likelihood}

\item{\code{doHeader}}{character string representing the header for DO to use. Null
value will prevent DO from being used in analysis.}

\item{\code{pCO2Header}}{Character string representing the header for pCO2 to use. Null
value will prevent pCO2 from being used in analysis.}

\item{\code{parHeader}}{Character string representing the header for PAR to use.}

\item{\code{tempHeader}}{Character string representing the header for temperature to use.}

\item{\code{airPressureHeader}}{Character string representing the header for air pressure to use.}

\item{\code{outputFile}}{The name of the output file.}

\item{\code{staticCO2Air}}{The pCO2 in the air.}

\item{\code{staticAlkalinity}}{The alkalinity.}

\item{\code{co2AirHeader}}{Character string representing the header for co2 partial pressure in air to use.}

\item{\code{alkalinityHeader}}{Character string representing the header for alkalinity to use.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TwoStationMetabMLE$new()}}
\item \href{#method-derive}{\code{TwoStationMetabMLE$derive()}}
\item \href{#method-clone}{\code{TwoStationMetabMLE$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabMLE$new(
  ...,
  initParams,
  parameterTranslatorGenerator = ParameterTranslatorMetab,
  isEstimated = c(TRUE, TRUE, TRUE),
  sdLikelihood,
  doHeader = "do",
  pCO2Header = "pCO2",
  parHeader = "par",
  tempHeader = "temp",
  airPressureHeader = "airPressure",
  staticAirPressure = NULL,
  outputFile = "results",
  staticCO2Air = NULL,
  staticAlkalinity = NULL,
  co2AirHeader = "co2Air",
  alkalinityHeader = "alkalinity"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments needed by the constructor of the super class}

\item{\code{initParams}}{The intial parameter values to use for the MLE algorithm}

\item{\code{parameterTranslatorGenerator}}{The R6 class generator for the translator used to convert the parameters 
provided into model input.
Default value is the R6 class generator ParameterTranslatorMetab.}

\item{\code{isEstimated}}{A vector of logical values indicating which parameters
to include in the estimation.}

\item{\code{sdLikelihood}}{The standard deviations used to calculate likelihood}

\item{\code{doHeader}}{Optional character string representing the header for DO to use. Null
value will prevent DO from being used in analysis.
Defaults to "do".}

\item{\code{pCO2Header}}{Optional character string representing the header for pCO2 to use. Null
value will prevent pCO2 from being used in analysis.
Defaults to "pCO2".}

\item{\code{parHeader}}{Optional character string representing the header for PAR to use. 
Defaults to "par".}

\item{\code{tempHeader}}{Optional character string representing the header for temperature to use. 
Defaults to "temp".}

\item{\code{airPressureHeader}}{Optional character string representing the header for air pressure to use.
Defaults to "airPressure"}

\item{\code{staticAirPressure}}{The air pressure.
Defaults to NULL, which will cause use of air pressure
from the signal.}

\item{\code{outputFile}}{The name of the output file.
Defaults to "results".}

\item{\code{staticCO2Air}}{The pCO2 in the air.
Defaults to NULL, which will cause use of pCO2
from the signal.}

\item{\code{staticAlkalinity}}{The alkalinity.
Defaults to NULL, which will cause use of alkalinity
from the signal.}

\item{\code{co2AirHeader}}{Optional character string representing the header for co2 partial pressure in air to use.
Defaults to "co2Air"}

\item{\code{alkalinityHeader}}{Optional character string representing the header for alkalinity to use.
Defaults to "alkalinity"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-derive"></a>}}
\if{latex}{\out{\hypertarget{method-derive}{}}}
\subsection{Method \code{derive()}}{
Uses a two-station model to infer whole-stream metabolism and
  gas exchange parameters from an upstream and downstream signal.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabMLE$derive(
  signalIn = NULL,
  signalOut = NULL,
  prevResults = NULL,
  path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signalIn}}{The input (upstream) signal on which the MLE inference is based}

\item{\code{signalOut}}{The output (downstream) signal on which the MLE inference is based}

\item{\code{prevResults}}{The results from the previous inference}

\item{\code{path}}{The path to where the results should be written}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The two-element list containint the results of the MLE.
  \itemize{
    \item objFunc: The objective function used for the MLE
    \item optimr: The results from optim performing the MLE
  }
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabMLE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
