% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoStationMetabPlotter.R
\name{TwoStationMetabPlotter}
\alias{TwoStationMetabPlotter}
\title{R6 class defining a plotter for a two station metabolism model}
\description{
Plots a signal of DO and pCO2 from a two station model of metabolism
}
\section{Super classes}{
\code{\link[disco:TransferFunctionSummarizer]{disco::TransferFunctionSummarizer}} -> \code{\link[disco:TransferFunctionPlotter]{disco::TransferFunctionPlotter}} -> \code{TwoStationMetabPlotter}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{timeTicks}}{Number of ticks on the time axis}

\item{\code{plotDelta}}{Logical value indicating if the change in concentration
should be plotted instead of the upstream/downstream
concentrations}

\item{\code{resultFile}}{Base file name for the results.}

\item{\code{results}}{The results object to be plotted}

\item{\code{subTitle}}{Subtitle for headings on plots, will appear over right
panel on legend.}

\item{\code{format}}{Data frame aggregating the formats for all the plots}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TwoStationMetabPlotter$new()}}
\item \href{#method-open}{\code{TwoStationMetabPlotter$open()}}
\item \href{#method-summarize}{\code{TwoStationMetabPlotter$summarize()}}
\item \href{#method-clone}{\code{TwoStationMetabPlotter$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="disco" data-topic="TransferFunctionPlotter" data-id="close">}\href{../../disco/html/TransferFunctionPlotter.html#method-close}{\code{disco::TransferFunctionPlotter$close()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabPlotter$new(
  ...,
  timeTicks,
  plotDelta = FALSE,
  resultFile = "results",
  subTitle = "",
  format.doObsUp = list(name = "[DO] observed upstream", lty = NA, pch = 46, col =
    "black", lwd = 1),
  format.doObsDown = list(name = "[DO] observed downstream", lty = NA, pch = 176, col =
    "black", lwd = 1),
  format.doPred = list(name = "[DO] modeled downstream", lty = "solid", pch = NA, col =
    "black", lwd = 2),
  format.parUp = list(name = "PAR upstream", lty = NA, pch = 46, col = "red", lwd = 1),
  format.parDown = list(name = "PAR downstream", lty = NA, pch = 176, col = "red", lwd
    = 1),
  format.pCO2ObsUp = list(name = "pCO2 observed upstream", lty = NA, pch = 46, col =
    "green3", lwd = 1),
  format.pCO2ObsDown = list(name = "pCO2 observed downstream", lty = NA, pch = 176, col
    = "green3", lwd = 1),
  format.pCO2Pred = list(name = "pCO2 modeled downstream", lty = "solid", pch = NA, col
    = "green3", lwd = 2),
  format.tempUp = list(name = "Temperature upstream", lty = NA, pch = 46, col =
    "black", lwd = 1),
  format.tempDown = list(name = "Temperature downstream", lty = NA, pch = 176, col =
    "black", lwd = 1),
  format.doSatDown = list(name = "[DO] saturated downstream", lty = "dotted", pch = NA,
    col = "black", lwd = 1)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to the constructor of the super class}

\item{\code{timeTicks}}{Number of ticks on the time axis}

\item{\code{plotDelta}}{Logical value indicating if the change in concentration
should be plotted instead of the upstream/downstream
concentrations}

\item{\code{resultFile}}{Base file name for the results.
Defaults to "results".}

\item{\code{subTitle}}{Subtitle for headings on plots, will appear over right
panel on legend.
Defaults to an empty string (blank).}

\item{\code{format.doObsUp}}{Format for upstream DO observations.
See defaults in usage.}

\item{\code{format.doObsDown}}{Format for downstream DO observations.
See defaults in usage.}

\item{\code{format.doPred}}{Format for downstream DO predictions
See defaults in usage.}

\item{\code{format.parUp}}{Format for upstream PAR
See defaults in usage.}

\item{\code{format.parDown}}{Format for downstream PAR
See defaults in usage.}

\item{\code{format.pCO2ObsUp}}{Format for upstream pCO2 observations.
See defaults in usage.}

\item{\code{format.pCO2ObsDown}}{Format for downstream pCO2 observations.
See defaults in usage.}

\item{\code{format.pCO2Pred}}{Format for pCO2 predictions
See defaults in usage.}

\item{\code{format.tempUp}}{Format for upstream temperature
See defaults in usage.}

\item{\code{format.tempDown}}{Format for downstream temperature
See defaults in usage.}

\item{\code{format.doSatDown}}{Format for downstream saturation DO concentrations
See defaults in usage.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-open"></a>}}
\if{latex}{\out{\hypertarget{method-open}{}}}
\subsection{Method \code{open()}}{
Opens the pdf for writing the plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabPlotter$open(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path where the pdf of figures should be written}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-summarize}{}}}
\subsection{Method \code{summarize()}}{
Summarize the data in a two panel visualization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabPlotter$summarize(
  signalIn = NULL,
  signalOut = NULL,
  outputPath = NULL,
  label,
  timeBounds
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signalIn}}{The input signal object that contains upstream observations}

\item{\code{signalOut}}{The output signal object that contains downstream observations}

\item{\code{outputPath}}{A path to results of a metabolism analysis.
Not necessary if summarizer does not need output.}

\item{\code{label}}{A label for the summary}

\item{\code{timeBounds}}{The temporal bounds on the summary}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoStationMetabPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
